[refactor] Controller refactor
